+++ cmd/src/campaigns_repositories.go
@@ -49,6 +49,11 @@ Examples:
 		client := cfg.apiClient(apiFlags, flagSet.Output())
 
 		svc := campaigns.NewService(&campaigns.ServiceOpts{Client: client})
+
+		if err := svc.DetermineFeatureFlags(ctx); err != nil {
+			return err
+		}
+
 		spec, _, err := svc.ParseCampaignSpec(specFile)
 		if err != nil {
 			return errors.Wrap(err, "parsing campaign spec")