+++ cmd/src/campaigns_validate.go
@@ -39,6 +39,7 @@ Examples:
 		defer specFile.Close()
 
 		svc := campaigns.NewService(&campaigns.ServiceOpts{})
+
 		spec, _, err := svc.ParseCampaignSpec(specFile)
 		if err != nil {
 			return errors.Wrap(err, "parsing campaign spec")